socket: add support for "wildcard" key
iptables had a -m socket --transparent didn't match sockets that are
bound to all addresses (e.g. 0.0.0.0 for ipv4, and ::0 for ipv6), which
could be overridden by using --nowildcard, in which case it did match
those as well.

The issue is that nftables never included the wildcard check, so in effect
it behaved like "iptables -m socket --transparent --nowildcard" with no
means to exclude wildcarded listeners.

This might be a problem, as a user-space process that binds to
0.0.0.0:<port> that enables IP_TRANSPARENT would effectively
intercept traffic going in _any_ direction on the specific port, whereas
in most cases, transparent proxies would only need this for one
specific address.

The solution is to add "socket wildcard" key to the nft_socket module, which
makes it possible to match on the wildcardness of a socket from
one's ruleset.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>